+++ config/nova-media-field.php
@@ -18,11 +18,20 @@ return [
 
     'collections' => [],
 
-    'storage_driver' => config('filesystems.default'),
+    'storage_driver' => env('MEDIA_LIBRARY_DRIVER', 'public'),
 
-    'storage_path' => 'public/media/',
+    'storage_path' => 'media/',
 
     'media_handler' => \OptimistDigital\MediaField\Classes\MediaHandler::class,
 
     'media_resource' => \OptimistDigital\MediaField\Media::class,
+
+    // Allows to rewrite core logic for custom needs. NB: Don't use it to modify json responses!
+    'media_model' => \OptimistDigital\MediaField\Models\Media::class,
+
+    // When enabled tries to find if file already exists and serve that instead of creating a duplicate entry
+    'resolve_duplicates' => true,
+
+    // Encoding quality for thumbnails and for source image
+    'quality' => 80,
 ];