+++ resources/js/components/FormField.vue
@@ -205,6 +205,7 @@ export default {
             if (this.updateFiles !== cb) cb();
           }
         }
+        this.updateFiles();
       }, 200)();
     },