+++ resources/js/components/IndexButton.vue
@@ -11,6 +11,7 @@
         :activeFile.sync="activeFile"
         :showUploadArea.sync="showUploadArea"
         @updateFiles="updateFiles"
+        @updateMedia="updateMedia"
         :loadingMediaFiles.sync="loadingMediaFiles"
         :selectedFiles.sync="selectedFiles"
       />
@@ -34,6 +35,14 @@ export default {
   mixins: [FormField, HandlesValidationErrors],
 
   props: ['field', 'onUploadFinished'],
+
+  mounted() {
+    window.addEventListener('open-media-library', this.openMediaBrowsingModal);
+  },
+
+  beforeDestroy() {
+    window.removeEventListener('open-media-library', this.openMediaBrowsingModal);
+  },
 
   data() {
     return {
@@ -71,6 +80,7 @@ export default {
 
     updateMedia: debounce(function () {
       this.onUploadFinished();
+      this.isModalOpen = false;
       this.$toasted.show('Files have been successfully uploaded!', { type: 'success' });
     }, 1000),