+++ src/Commands/RegenerateThumbnails.php
@@ -40,7 +40,8 @@ class RegenerateThumbnails extends Command
      */
     public function handle()
     {
-        $medias = Media::all();
+        $Media = config('nova-media-field.media_model');
+        $medias = $Media::all();
 
         /** @var MediaHandler $handler */
         $handler = app()->make(MediaHandler::class);